fix(web): fix sidebar tooltip pluralization
